Branch0001

+++ src/External/Restaurant.Persistence/Seeders/Misc/ImageSeeder.cs
@@ -56,7 +56,7 @@ namespace Restaurant.Persistence.Seeders.Misc
             public Guid Id { get; set; }
             public string? AltText { get; set; }
             public string? ContentType { get; set; }
-            public long FileSize { get; set; }
+            public decimal FileSize { get; set; }
             public bool IsPrimary { get; set; }
             public string? Url { get; set; }
             public string? StoragePath { get; set; }
